+++ docs/iscsi/storage-class.yaml
@@ -14,3 +14,4 @@ parameters:
   fsType: ext4 # Desired filesystem
   pool: A # Pool for volumes provisioning
   volPrefix: stx # Desired prefix for volume naming, an underscore is appended
+  storageProtocol: iscsi # The storage interface (iscsi, fc, sas) being used for storage i/o
