+++ example/storage-class.yaml
@@ -15,3 +15,4 @@ parameters:
   csi.storage.k8s.io/controller-expand-secret-namespace: default
   fsType: ext4 # Desired filesystem
   pool: A # Pool to use on the IQN to provision volumes
+  storageProtocol: iscsi # The storage interface (iscsi, fc, sas) being used for storage i/o
