+++ example/storageclass-example1.yaml
@@ -16,3 +16,4 @@ parameters:
   fsType: ext4 # Desired filesystem
   pool: A # Pool to use on the IQN to provision volumes
   volPrefix: csi # Desired prefix for volume naming, an underscore is appended
+  storageProtocol: iscsi # The storage interface (iscsi, fc, sas) being used for storage i/o
