+++ pkg/node/node.go
@@ -6,16 +6,13 @@ import (
 	"fmt"
 	"os"
 	"os/exec"
-	"regexp"
-	"strconv"
 	"strings"
-	"time"
 
 	"github.com/Seagate/csi-lib-iscsi/iscsi"
 	"github.com/Seagate/seagate-exos-x-csi/pkg/common"
+	"github.com/Seagate/seagate-exos-x-csi/pkg/storage"
 	"github.com/container-storage-interface/spec/lib/go/csi"
 	"github.com/golang/protobuf/ptypes/wrappers"
-	"github.com/pkg/errors"
 	"golang.org/x/sync/semaphore"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/codes"
@@ -135,219 +132,77 @@ func (node *Node) NodeGetCapabilities(ctx context.Context, req *csi.NodeGetCapab
 
 // NodePublishVolume mounts the volume mounted to the staging path to the target path
 func (node *Node) NodePublishVolume(ctx context.Context, req *csi.NodePublishVolumeRequest) (*csi.NodePublishVolumeResponse, error) {
-	if len(req.GetVolumeId()) == 0 {
-		return nil, status.Error(codes.InvalidArgument, "cannot publish volume with empty id")
-	}
-	if len(req.GetTargetPath()) == 0 {
-		return nil, status.Error(codes.InvalidArgument, "cannot publish volume at an empty path")
-	}
-	if req.GetVolumeCapability() == nil {
-		return nil, status.Error(codes.InvalidArgument, "cannot publish volume without capabilities")
-	}
 
-	klog.Infof("publishing volume %s", req.GetVolumeId())
-	klog.Infof("target path: %s", req.GetTargetPath())
+	klog.Infof("NodePublishVolume called with volume name %s", req.GetVolumeId())
 
-	iqn := req.GetVolumeContext()["iqn"]
-	portals := strings.Split(req.GetVolumeContext()["portals"], ",")
-	klog.Infof("iSCSI iqn: %s, portals: %v", iqn, portals)
+	storagePorotcol := req.GetVolumeContext()[common.StorageProtocolKey]
 
-	lun, _ := strconv.ParseInt(req.GetPublishContext()["lun"], 10, 32)
-	klog.Infof("LUN: %d", lun)
+	config := make(map[string]string)
+	config["iscsiInfoPath"] = node.getIscsiInfoPath(req.GetVolumeId())
+	klog.V(2).Infof("NodePublishVolume iscsiInfoPath (%v)", config["iscsiInfoPath"])
 
-	klog.Info("initiating ISCSI connection...")
-	targets := make([]iscsi.TargetInfo, 0)
-	for _, portal := range portals {
-		if portal != "" {
-			klog.V(1).Infof("-- add iqn (%v) portal (%v)", iqn, portal)
-			targets = append(targets, iscsi.TargetInfo{
-				Iqn:    iqn,
-				Portal: portal,
-			})
-		}
-	}
-	connector := iscsi.Connector{
-		Targets:     targets,
-		Lun:         int32(lun),
-		DoDiscovery: true,
+	// Get storage handler
+	storageNode, err := storage.NewStorageNode(storagePorotcol, config)
+	if storageNode != nil {
+		return storageNode.NodePublishVolume(ctx, req)
 	}
 
-	path, err := iscsi.Connect(&connector)
-	if err != nil {
-		return nil, status.Error(codes.Unavailable, err.Error())
-	}
-	klog.Infof("attached device at %s", path)
-
-	if connector.Multipath {
-		klog.Info("device is using multipath")
-	} else {
-		klog.Info("device is NOT using multipath")
-	}
-
-	fsType := req.GetVolumeContext()[common.FsTypeConfigKey]
-	err = ensureFsType(fsType, path)
-	if err != nil {
-		return nil, status.Error(codes.Internal, err.Error())
-	}
-
-	if err = checkFs(path); err != nil {
-		return nil, status.Errorf(codes.DataLoss, "filesystem seems to be corrupted: %v", err)
-	}
-
-	out, err := exec.Command("findmnt", "--output", "TARGET", "--noheadings", path).Output()
-	mountpoints := strings.Split(strings.Trim(string(out), "\n"), "\n")
-	if err != nil || len(mountpoints) == 0 {
-		klog.Infof("mounting volume at %s", req.GetTargetPath())
-		os.Mkdir(req.GetTargetPath(), 00755)
-		out, err = exec.Command("mount", "-t", fsType, path, req.GetTargetPath()).CombinedOutput()
-		if err != nil {
-			return nil, status.Error(codes.Internal, string(out))
-		}
-	} else if len(mountpoints) == 1 {
-		if mountpoints[0] == req.GetTargetPath() {
-			klog.Infof("volume %s already mounted", req.GetTargetPath())
-		} else {
-			errStr := fmt.Sprintf("device has already been mounted somewhere else (%s instead of %s), please unmount first", mountpoints[0], req.GetTargetPath())
-			return nil, status.Error(codes.Internal, errStr)
-		}
-	} else if len(mountpoints) > 1 {
-		return nil, errors.New("device has already been mounted in several locations, please unmount first")
-	}
-
-	iscsiInfoPath := node.getIscsiInfoPath(req.GetVolumeId())
-	klog.Infof("saving ISCSI connection info in %s", iscsiInfoPath)
-	err = iscsi.PersistConnector(&connector, iscsiInfoPath)
-	if err != nil {
-		return nil, status.Error(codes.Internal, err.Error())
-	}
-
-	klog.Infof("successfully mounted volume at %s", req.GetTargetPath())
-	return &csi.NodePublishVolumeResponse{}, nil
+	klog.Errorf("NodePublishVolume error for storage protocol (%v): %v", storagePorotcol, err)
+	return nil, status.Errorf(codes.Internal, "Unable to process for storage protocol (%v)", storagePorotcol)
 }
 
 // NodeUnpublishVolume unmounts the volume from the target path
 func (node *Node) NodeUnpublishVolume(ctx context.Context, req *csi.NodeUnpublishVolumeRequest) (*csi.NodeUnpublishVolumeResponse, error) {
-	if len(req.GetVolumeId()) == 0 {
-		return nil, status.Error(codes.InvalidArgument, "cannot unpublish volume with an empty volume id")
-	}
-	if len(req.GetTargetPath()) == 0 {
-		return nil, status.Error(codes.InvalidArgument, "cannot unpublish volume with an empty target path")
-	}
 
-	klog.Infof("unpublishing volume %s at target path %s", req.GetVolumeId(), req.GetTargetPath())
+	klog.Infof("NodeUnpublishVolume volume %s at target path %s", req.GetVolumeId(), req.GetTargetPath())
 
-	_, err := os.Stat(req.GetTargetPath())
-	if err == nil {
-		klog.Infof("unmounting volume at %s", req.GetTargetPath())
-		out, err := exec.Command("mountpoint", req.GetTargetPath()).CombinedOutput()
-		if err == nil {
-			out, err := exec.Command("umount", "-l", req.GetTargetPath()).CombinedOutput()
-			if err != nil {
-				return nil, status.Error(codes.Internal, string(out))
-			}
-		} else {
-			klog.Warningf("assuming that volume is already unmounted: %s", out)
-		}
+	// TODO: This csi request message does not provide any coontext for passing the storage protocol.
+	// Need to find a way to track the storage call per VolumeId, or through Context
+	// One solution discovered is to use VolumeId$$StorageProtocol - adding storage protocol string to VolumeId
+	storagePorotcol := "iscsi"
 
-		err = os.Remove(req.GetTargetPath())
-		if err != nil && !os.IsNotExist(err) {
-			return nil, status.Error(codes.Internal, err.Error())
-		}
-	} else {
-		klog.Warningf("assuming that volume is already unmounted: %v", err)
+	config := make(map[string]string)
+	config["iscsiInfoPath"] = node.getIscsiInfoPath(req.GetVolumeId())
+	klog.V(2).Infof("NodeUnpublishVolume iscsiInfoPath (%v)", config["iscsiInfoPath"])
+
+	// Get storage handler
+	storageNode, err := storage.NewStorageNode(storagePorotcol, config)
+	if storageNode != nil {
+		return storageNode.NodeUnpublishVolume(ctx, req)
 	}
 
-	iscsiInfoPath := node.getIscsiInfoPath(req.GetVolumeId())
-	klog.Infof("loading ISCSI connection info from %s", iscsiInfoPath)
-	connector, err := iscsi.GetConnectorFromFile(iscsiInfoPath)
-	if err != nil {
-		if os.IsNotExist(err) {
-			klog.Warning(errors.Wrap(err, "assuming that ISCSI connection is already closed"))
-			return &csi.NodeUnpublishVolumeResponse{}, nil
-		}
-		return nil, status.Error(codes.Internal, err.Error())
-	}
-	klog.Infof("connector.DevicePath (%s)", connector.DevicePath)
-
-	if isVolumeInUse(connector.DevicePath) {
-		klog.Info("volume is still in use on the node, thus it will not be detached")
-		return &csi.NodeUnpublishVolumeResponse{}, nil
-	}
-
-	_, err = os.Stat(connector.DevicePath)
-	if err != nil && os.IsNotExist(err) {
-		klog.Warningf("assuming that volume is already disconnected: %s", err)
-		return &csi.NodeUnpublishVolumeResponse{}, nil
-	}
-
-	if err = checkFs(connector.DevicePath); err != nil {
-		return nil, status.Errorf(codes.DataLoss, "Filesystem seems to be corrupted: %v", err)
-	}
-
-	klog.Info("DisconnectVolume, detaching ISCSI device")
-	err = iscsi.DisconnectVolume(*connector)
-	if err != nil {
-		return nil, err
-	}
-
-	klog.Infof("deleting ISCSI connection info file %s", iscsiInfoPath)
-	os.Remove(iscsiInfoPath)
-
-	klog.Info("successfully detached ISCSI device")
-	return &csi.NodeUnpublishVolumeResponse{}, nil
+	klog.Errorf("NodeUnpublishVolume error for storage protocol (%v): %v", storagePorotcol, err)
+	return nil, status.Errorf(codes.Internal, "Unable to process for storage protocol (%v)", storagePorotcol)
 }
 
 // NodeExpandVolume finalizes volume expansion on the node
 func (node *Node) NodeExpandVolume(ctx context.Context, req *csi.NodeExpandVolumeRequest) (*csi.NodeExpandVolumeResponse, error) {
 
-	volumeid := req.GetVolumeId()
-	volumepath := req.GetVolumePath()
-	klog.V(2).Infof("NodeExpandVolume: VolumeId=%v,  VolumePath=%v", volumeid, volumepath)
+	klog.Infof("NodeExpandVolume volume %s at volume path %s", req.GetVolumeId(), req.GetVolumePath())
 
-	if len(volumeid) == 0 {
-		return nil, status.Error(codes.InvalidArgument, fmt.Sprintf("node expand volume requires volume id"))
+	// TODO: This csi request message does not provide any coontext for passing the storage protocol.
+	// Need to find a way to track the storage call per VolumeId, or through Context
+	// One solution discovered is to use VolumeId$$StorageProtocol - adding storage protocol string to VolumeId
+	storagePorotcol := "iscsi"
+
+	config := make(map[string]string)
+	config["iscsiInfoPath"] = node.getIscsiInfoPath(req.GetVolumeId())
+	klog.V(2).Infof("NodeExpandVolume iscsiInfoPath (%v)", config["iscsiInfoPath"])
+
+	// Get storage handler
+	storageNode, err := storage.NewStorageNode(storagePorotcol, config)
+	if storageNode != nil {
+		return storageNode.NodeExpandVolume(ctx, req)
 	}
 
-	if len(volumepath) == 0 {
-		return nil, status.Error(codes.InvalidArgument, fmt.Sprintf("node expand volume requires volume path"))
-	}
-
-	iscsiInfoPath := node.getIscsiInfoPath(volumeid)
-	connector, err := iscsi.GetConnectorFromFile(iscsiInfoPath)
-	klog.V(3).Infof("GetConnectorFromFile(%s) connector: %v, err: %v", volumeid, connector, err)
-
-	if err != nil {
-		return nil, status.Error(codes.NotFound, fmt.Sprintf("node expand volume path not found for volume id (%s)", volumeid))
-	}
-
-	// TODO: Is a rescan needed - rescan a scsi device by writing 1 in /sys/class/scsi_device/h:c:t:l/device/rescan
-	// for i := range connector.Devices {
-	// 	connector.Devices[i].Rescan()
-	// }
-
-	if connector.Multipath {
-		klog.V(2).Info("device is using multipath")
-		if err := iscsi.ResizeMultipathDevice(connector.DevicePath); err != nil {
-			return nil, err
-		}
-	} else {
-		klog.V(2).Info("device is NOT using multipath")
-	}
-
-	klog.Infof("expanding filesystem using resize2fs on device %s", connector.DevicePath)
-	output, err := exec.Command("resize2fs", connector.DevicePath).CombinedOutput()
-	if err != nil {
-		klog.V(2).Info("could not resize filesystem: %v", output)
-		return nil, fmt.Errorf("could not resize filesystem: %v", output)
-	}
-
-	return &csi.NodeExpandVolumeResponse{}, nil
+	klog.Errorf("NodeExpandVolume error for storage protocol (%v): %v", storagePorotcol, err)
+	return nil, status.Errorf(codes.Internal, "Unable to process for storage protocol (%v)", storagePorotcol)
 }
 
 // NodeGetVolumeStats return info about a given volume
 // Will not be called as the plugin does not have the GET_VOLUME_STATS capability
 func (node *Node) NodeGetVolumeStats(ctx context.Context, req *csi.NodeGetVolumeStatsRequest) (*csi.NodeGetVolumeStatsResponse, error) {
-	return nil, status.Error(codes.Unimplemented, "NodeGetVolumeStats is unimplemented and should not be called")
+	return nil, status.Error(codes.Unimplemented, "NodeGetVolumeStats is not implemented")
 }
 
 // NodeStageVolume mounts the volume to a staging path on the node. This is
@@ -356,13 +211,13 @@ func (node *Node) NodeGetVolumeStats(ctx context.Context, req *csi.NodeGetVolume
 // mount it to the appropriate path
 // Will not be called as the plugin does not have the STAGE_UNSTAGE_VOLUME capability
 func (node *Node) NodeStageVolume(ctx context.Context, req *csi.NodeStageVolumeRequest) (*csi.NodeStageVolumeResponse, error) {
-	return nil, status.Error(codes.Unimplemented, "NodeStageVolume is unimplemented and should not be called")
+	return nil, status.Error(codes.Unimplemented, "NodeStageVolume is not implemented")
 }
 
 // NodeUnstageVolume unstages the volume from the staging path
 // Will not be called as the plugin does not have the STAGE_UNSTAGE_VOLUME capability
 func (node *Node) NodeUnstageVolume(ctx context.Context, req *csi.NodeUnstageVolumeRequest) (*csi.NodeUnstageVolumeResponse, error) {
-	return nil, status.Error(codes.Unimplemented, "NodeUnstageVolume is unimplemented and should not be called")
+	return nil, status.Error(codes.Unimplemented, "NodeUnstageVolume is not implemented")
 }
 
 // Probe returns the health and readiness of the plugin
@@ -371,10 +226,12 @@ func (node *Node) Probe(ctx context.Context, req *csi.ProbeRequest) (*csi.ProbeR
 	return &csi.ProbeResponse{Ready: &wrappers.BoolValue{Value: true}}, nil
 }
 
+// getIscsiInfoPath
 func (node *Node) getIscsiInfoPath(volumeID string) string {
 	return fmt.Sprintf("%s/iscsi-%s.json", node.runPath, volumeID)
 }
 
+// checkHostBinary: Determine if a binary image is installed or not
 func checkHostBinary(name string) error {
 	if path, err := exec.LookPath(name); err != nil {
 		return fmt.Errorf("binary %q not found", name)
@@ -385,89 +242,7 @@ func checkHostBinary(name string) error {
 	return nil
 }
 
-func checkFs(path string) error {
-	klog.Infof("Checking filesystem at %s", path)
-	if out, err := exec.Command("e2fsck", "-n", path).CombinedOutput(); err != nil {
-		return errors.New(string(out))
-	}
-	return nil
-}
-
-func findDeviceFormat(device string) (string, error) {
-	klog.V(2).Infof("Trying to find filesystem format on device %q", device)
-
-	ctx, cancel := context.WithTimeout(context.Background(), 2*time.Second)
-	defer cancel()
-	output, err := exec.CommandContext(ctx, "blkid",
-		"-p",
-		"-s", "TYPE",
-		"-s", "PTTYPE",
-		"-o", "export",
-		device).CombinedOutput()
-
-	if ctx.Err() == context.DeadlineExceeded {
-		err = errors.New("command timed out after 2 seconds")
-	}
-
-	klog.V(2).Infof("blkid output: %q, err=%v", output, err)
-
-	if err != nil {
-		// blkid exit with code 2 if the specified token (TYPE/PTTYPE, etc) could not be found or if device could not be identified.
-		if exit, ok := err.(*exec.ExitError); ok && exit.ExitCode() == 2 {
-			klog.V(2).Infof("Device seems to be is unformatted (%v)", err)
-			return "", nil
-		}
-		return "", fmt.Errorf("could not not find format for device %q (%v)", device, err)
-	}
-
-	re := regexp.MustCompile(`([A-Z]+)="?([^"\n]+)"?`) // Handles alpine and debian outputs
-	matches := re.FindAllSubmatch(output, -1)
-
-	var filesystemType, partitionType string
-	for _, match := range matches {
-		if len(match) != 3 {
-			return "", fmt.Errorf("invalid blkid output: %s", output)
-		}
-		key := string(match[1])
-		value := string(match[2])
-
-		if key == "TYPE" {
-			filesystemType = value
-		} else if key == "PTTYPE" {
-			partitionType = value
-		}
-	}
-
-	if partitionType != "" {
-		klog.V(2).Infof("Device %q seems to have a partition table type: %s", partitionType)
-		return "OTHER/PARTITIONS", nil
-	}
-
-	return filesystemType, nil
-}
-
-func ensureFsType(fsType string, disk string) error {
-	currentFsType, err := findDeviceFormat(disk)
-	if err != nil {
-		return err
-	}
-
-	klog.V(1).Infof("Detected filesystem: %q", currentFsType)
-	if currentFsType != fsType {
-		if currentFsType != "" {
-			return fmt.Errorf("Could not create %s filesystem on device %s since it already has one (%s)", fsType, disk, currentFsType)
-		}
-
-		klog.Infof("Creating %s filesystem on device %s", fsType, disk)
-		out, err := exec.Command(fmt.Sprintf("mkfs.%s", fsType), disk).CombinedOutput()
-		if err != nil {
-			return errors.New(string(out))
-		}
-	}
-
-	return nil
-}
-
+// readInitiatorName: Extract the initiaotr name from /etc/iscsi file
 func readInitiatorName() (string, error) {
 	initiatorNameFilePath := "/etc/iscsi/initiatorname.iscsi"
 	file, err := os.Open(initiatorNameFilePath)
@@ -491,16 +266,4 @@ func readInitiatorName() (string, error) {
 	}
 
 	return "", fmt.Errorf("InitiatorName key is missing from %s", initiatorNameFilePath)
-}
-
-// isVolumeInUse: Use findmnt to determine if the devie path is mounted or not.
-func isVolumeInUse(devicePath string) bool {
-	_, err := exec.Command("findmnt", devicePath).CombinedOutput()
-	klog.Infof("isVolumeInUse: findmnt %s, err=%v", devicePath, err)
-	if err != nil {
-		if _, ok := err.(*exec.ExitError); ok {
-			return false
-		}
-	}
-	return true
 }
